+++ rcqt.so/src/handlerCreator.cpp
@@ -31,10 +31,10 @@ PackageHandler* handlerCreator::getPackageHandler(const std::string& pkg){
   auto osName = getOS();
   PackageHandler* lptr = nullptr;
 
-  if(OSType::Ubuntu == osName){
+  if(OSType::Ubuntu == osName) {
     lptr = new PackageHandlerDeb{pkg};
   }
-  else if (OSType::Centos == osName || OSType::RHEL == osName) {
+  else if (OSType::Centos == osName || OSType::RHEL == osName || OSType::Oracle == osName) {
     lptr = new PackageHandlerRpm{pkg};
   }
   else if (OSType::SLES == osName) {
@@ -51,7 +51,7 @@ PackageHandler* handlerCreator::getPackageHandler(){
   if(OSType::Ubuntu == osName){
     lptr = new PackageHandlerDeb{};
   }
-  else if (OSType::Centos == osName || OSType::RHEL == osName) {
+  else if (OSType::Centos == osName || OSType::RHEL == osName || OSType::Oracle == osName) {
     lptr = new PackageHandlerRpm{};
   }
   else if (OSType::SLES == osName) {
@@ -59,3 +59,4 @@ PackageHandler* handlerCreator::getPackageHandler(){
   }
   return lptr;
 }
+