+++ CHANGELOG.md
@@ -7,6 +7,9 @@ Full documentation for RVS is available at [ROCmValidationSuite.Readme](https://
 ### Changed
 - Updated pebb & pbqt logs to include PCI BDF.
 
+### Added
+- Support data types (BF16 and FP8) based GEMM operations in GPU Stress Test (GST) module.
+
 ## RVS for ROCm 6.0
 
 ### Added